refactor: remove torch_dtype parameter from model initialization
The torch_dtype parameter was removed from the BartForConditionalGeneration initialization as it is not necessary for the model to function correctly. This simplifies the code and avoids potential compatibility issues with different hardware setups.

+++ train.py
@@ -84,7 +84,7 @@ class ArgsBase():
 class KoBARTSummaryModel(nn.Module):
     def __init__(self, tokenizer):
         super().__init__()
-        self.model = BartForConditionalGeneration.from_pretrained('gogamza/kobart-base-v2', torch_dtype=torch.bfloat16)
+        self.model = BartForConditionalGeneration.from_pretrained('gogamza/kobart-base-v2')
         self.model.resize_token_embeddings(len(tokenizer))
         self.pad_token_id = tokenizer.pad_token_id
